Crate ym_file_parser

YM-file parser and player helper.

This format was designed by Leonard/OXYGENE for his AY-emulator StSound.

YM-files are distributed as compressed LHA archives.

This library can help uncompress, parse the YM-files, and produce the AY/YM register changes for the players.

The following YM-file types are supported: YM2!, YM3!, YM3B, YM4!, YM5! and YM6!.

All special effects described by Leonard are being recognized. Here is the description of how those special effects and AY/YM register data are being encoded in frames.

Modules

effects

Special effects.

flags

YmSong related flags.

Structs

YmFrame

This type represent the state of the AY/YM chipset registers and contain additional information about special effects.

YmSong

The YM music file.

Enums

YmVersion

Constants

MAX_DD_SAMPLES

MFP_TIMER_FREQUENCY

Functions

parse_file

Attempts to parse an YM-file that can be either compressed or uncompressed, from the given file path.